repolint.json: remove bb-file-license-exist check
It is a common practice in OE/Yocto for recipes to include other files
(e.g. .inc) that contains the LICENSE declaration (e.g. recipes for
multiple versions of the same software component), so this check will
quite often fail.

Since LICENSE is not the license of the bb file itself, but for the
software the recipe is providing build instructions for, we can rely on
OE to verify that there is a valid declaration instead of having it as
part of repolinter.

Fixes #33.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>