repolint.json: allow community copyright owners
As we are opening up several repositories to accept contributions from
external developers, remove the check that enforces Copyright to be set
to Qualcomm or Linux Foundation only, as it can be assigned to other
companies and users. Year enforcement is also not required, so removed.

This changes to just force the check to verify that there is a Copyright
entry instead.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>